Add files via upload
A Simple Linear Regression for Breast Cancer Detection
Problem Statement :
Use a simple linear regression for breast cancer detection.

Dataset link : http://archive.ics.uci.edu/ml/datasets/Breast+Cancer+Wisconsin+%28Diagnostic%29Links

The input matrix is called P, each row of which corresponds to different measurements of a patient’s tumor cell sample. The corresponding element in the output vector T is -1 if the cell was determined to be benign and +1 otherwise.

Task 1:
(a) Plot the ROC curve and
(b) Calculate TPRs at FPR of 5% and 10%, alongside the corresponding thresholds. Use the whole dataset.

Task 2 :
(a) Use the first 2/3 or 66.67% of the data for training and 1/3 or 33.33% of the data for testing.
(b) Plot the ROC curve
(c) Calculate TPRs at FPR of 5% and 10%, alongside the corresponding thresholds. Use the whole dataset.
(d) Apply training thresholds to the test data and calculate TPR/FPR pairs for the test subset, use the same decision thresholds obtained for training's 5% and 10% FPRs